Move base package to com.github.azulaloi
This will prevent conflicts in the future, as well as allow you to
publish this repository and artifacts in a maven repository
