+++ gradle.properties
@@ -1,6 +1,6 @@
 mod_version = 1.0
-group = com.yourname.modid
-archivesBaseName = modid
+group = com.github.azulaloi.dtnatura
+archivesBaseName = dtnatura
 
 minecraft_version = 1.12.2
 forge_version = 14.23.4.2705
