+++ src/main/java/com/github/azulaloi/dtnatura/proxy/CommonProxy.java
@@ -1,0 +1,22 @@
+package com.github.azulaloi.dtnatura.proxy;
+
+import net.minecraftforge.fml.common.event.FMLInitializationEvent;
+import net.minecraftforge.fml.common.event.FMLPostInitializationEvent;
+import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
+
+public class CommonProxy {
+
+    public void preInit(FMLPreInitializationEvent event) {
+
+    }
+
+    public void init(FMLInitializationEvent event) {
+
+    }
+
+    public void postInit(FMLPostInitializationEvent event) {
+
+    }
+
+
+}
