--- src/main/java/dtnatura/trees/overworld/TreeEucalyptus.java
@@ -1,67 +1,0 @@
-package dtnatura.trees.overworld;
-
-import com.ferreusveritas.dynamictrees.blocks.BlockDynamicSapling;
-import com.ferreusveritas.dynamictrees.trees.Species;
-import com.ferreusveritas.dynamictrees.trees.TreeFamily;
-import com.progwml6.natura.overworld.NaturaOverworld;
-import com.progwml6.natura.overworld.block.logs.BlockOverworldLog;
-import com.progwml6.natura.overworld.block.logs.BlockOverworldLog2;
-import dtnatura.DynamicTreesNatura;
-import dtnatura.ModBlocks;
-import net.minecraft.block.Block;
-import net.minecraft.block.state.IBlockState;
-import net.minecraft.init.Biomes;
-import net.minecraft.item.ItemStack;
-import net.minecraft.util.ResourceLocation;
-import net.minecraft.util.math.MathHelper;
-import net.minecraft.world.biome.Biome;
-import net.minecraftforge.common.BiomeDictionary;
-
-import java.util.List;
-
-public class TreeEucalyptus extends TreeFamily {
-    public class SpeciesEucalyptus extends Species {
-
-        SpeciesEucalyptus (TreeFamily treeFamily) {
-            super(treeFamily.getName(), treeFamily, ModBlocks.eucalyptusLeavesProperties);
-
-            setBasicGrowingParameters(0.3f, 12.0f, upProbability, lowestBranchHeight, 8f);
-
-            setDynamicSapling(new BlockDynamicSapling("eucalyptussapling").getDefaultState());
-
-            envFactor(BiomeDictionary.Type.FOREST, 1.5f);
-
-            generateSeed();
-            setupStandardSeedDropping();
-        }
-
-        @Override
-        public boolean isBiomePerfect(Biome biome) { return isOneOfBiomes(biome, Biomes.PLAINS); }
-    }
-
-    public TreeEucalyptus() {
-        super(new ResourceLocation(DynamicTreesNatura.MODID, "eucalyptus"));
-
-        IBlockState primLog = NaturaOverworld.overworldLog2.getDefaultState().withProperty(BlockOverworldLog2.TYPE, BlockOverworldLog2.LogType.EUCALYPTUS);
-
-        setPrimitiveLog(primLog);
-
-        ModBlocks.eucalyptusLeavesProperties.setTree(this);
-    }
-
-    @Override
-    public ItemStack getPrimitiveLogItemStack(int qty) {
-        return new ItemStack(NaturaOverworld.overworldLog2, MathHelper.clamp(qty, 0, 64), 1);
-    }
-
-    @Override
-    public void createSpecies() {
-        setCommonSpecies(new SpeciesEucalyptus(this));
-    }
-
-    @Override
-    public List<Block> getRegisterableBlocks(List<Block> blockList) {
-        blockList.add(getCommonSpecies().getDynamicSapling().getBlock());
-        return super.getRegisterableBlocks(blockList);
-    }
-}
