+++ src/main/resources/pack.mcmeta
@@ -1,6 +1,6 @@
 {
     "pack": {
-        "description": "dtnatura resources",
+        "description": "com.github.azulaloi.dtnatura resources",
         "pack_format": 3
     }
 }
